implemented option to start less stable mode
